net: Change ticcmd_t.consistancy back to a short.
Vanilla Doom (plus all other derived games) actually use a 16-bit
consistency field, but for some reason back in 2006 I decided to change
this to an 8-bit field. This isn't going to work if we want to talk the
vanilla networking protocol.

To remain protocol-compatible with other Chocolate Doom versions, keep the
8-bit wire format, but allow consistency values where the top byte is
missing.

This reverts commit ce765250b6265bba27174a2339fd14cea2bb0925.

+++ src/d_ticcmd.h
@@ -36,15 +36,13 @@ typedef struct
     short angleturn;            // <<16 for angle delta
     byte chatchar;
     byte buttons;
-    // villsa [STRIFE] according to the asm,
-    // consistancy is a short, not a byte
-    byte consistancy;           // checks for net game
+    short consistancy;           // checks for net game
 
     // villsa - Strife specific:
 
     byte buttons2;
     int inventory;
-   
+
     // Heretic/Hexen specific:
 
     byte lookfly;               // look/fly up/down/centering
